Moving Feed / Json strategies.
They are no longer listener aggregates. They now implement getFieldValueParts to obtain AcceptFieldValueParts used in matching.
